サンプルモデルを修正, 学習をfit関数を用いるように修正

+++ utils/read_data.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import os
-
+import sys
 def read_horse_csv(path):
     df = pd.DataFrame()
     for file in os.listdir(path):
@@ -9,7 +9,7 @@ def read_horse_csv(path):
             df = pd.concat([df, pd.read_csv(path  + file)])
     df['race_id'] = df['race_id'].astype(str)
     df = df.query('not race_id.str.contains("\(", na=False)', engine='python')
-    df['race_id'] = df['race_id'].astype(int)
+    df['race_id'] = df['race_id'].astype(float)
     df = df.reset_index(drop=True)
     return df
 
@@ -22,6 +22,6 @@ def read_race_csv(path):
             df = pd.concat([df, pd.read_csv(path  + file)])
     df['race_id'] = df['race_id'].astype(str)
     df = df.query('not race_id.str.contains("\(", na=False)', engine='python')
-    df['race_id'] = df['race_id'].astype(int)
+    df['race_id'] = df['race_id'].astype(float)
     df = df.reset_index(drop=True)
     return df
